Add files via upload
An array in JavaScript is a type of object that can store multiple values in a single variable. Arrays are zero-indexed, meaning that the first element in an array is at index 0, and the last element is at index length - 1. Arrays can store any type of data, including numbers, strings, objects, and even other arrays.